// sanity.types.ts
export type SeasonReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "season";
};

export type Lookbook = {
  _id: string;
  _type: "lookbook";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  slug?: Slug;
  season?: SeasonReference;
  category?: "fastFashion" | "ready-to-wear" | "couture";
  description?: string;
  image?: ImageReference;
  gallery?: Array<
    {
      _key: string;
    } & ImageReference
  >;
  publishedAt?: string;
};

export type ImageReference = {
  _type: "imageReference";
  url?: string;
  alt?: string;
  caption?: string;
};

export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};

export type Project = {
  _id: string;
  _type: "project";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  title?: string;
  slug?: Slug;
  season?: SeasonReference;
  category?: "fastFashion" | "ready-to-wear" | "couture";
  description?: string;
  image?: ImageReference;
  gallery?: Array<
    {
      _key: string;
    } & ImageReference
  >;
  publishedAt?: string;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;